Рефакторинг2
